chore: make aggregator abstraction compatible with latest changes
