+++ tests/src/aggregator/mod.rs
@@ -1,2 +1,3 @@
-#[cfg(test)]
+#![cfg(test)]
+pub mod aggregator_tests;
 pub mod circuit_tests;
